Move preload component to playback.js
