+++ frontend/src/components/Playback/Autoplay.js
@@ -2,24 +2,15 @@ import React, { useRef, useState, useEffect } from "react";
 
 import * as audio from "../../util/audio";
 import * as webAudio from "../../util/webAudio";
-import { MEDIA_ROOT } from "../../config";
 
 import Circle from "../Circle/Circle";
 import ListenCircle from "../ListenCircle/ListenCircle";
-import Preload from "../Preload/Preload";
-
-const PRELOAD = "PRELOAD";
-const RECOGNIZE = "RECOGNIZE";
-
 
 const AutoPlay = ({instruction, preloadMessage, onPreloadReady, playConfig, sections, time, startedPlaying, finishedPlaying, responseTime, className=''}) => {
     // player state
-    const [state, setState] = useState({ view: RECOGNIZE });
+    
     const running = useRef(playConfig.auto_play);
-    const setView = (view, data = {}) => {
-        setState({ view, ...data });
-    }
-
+    
     const section = sections[0];
 
     const onCircleTimerTick = (t) => {
@@ -27,90 +18,61 @@ const AutoPlay = ({instruction, preloadMessage, onPreloadReady, playConfig, sect
     };
 
     // Handle view logic
-    useEffect(() => {
-        switch (state.view) {
-            case RECOGNIZE:
-                let latency = 0;
-                // Play audio at start time            
-                if (!playConfig.mute) {
-                    if (playConfig.play_method === 'BUFFER' && !playConfig.external_audio) {
-                        console.log('Autoplay buffer');
-                        latency = webAudio.getTotalLatency();
-                        webAudio.playBufferFrom(section.id, 0, Math.max(0, playConfig.playhead));
-                    } else {
-                        console.log('Autoplay HTML audio')
-                        audio.playFrom(Math.max(0, playConfig.playhead));
-                    }                    
-                }
-                setTimeout(startedPlaying(), latency);
-                break;
-            default:
-            // nothing
+    useEffect(() => {        
+        let latency = 0;
+        // Play audio at start time            
+        if (!playConfig.mute) {
+            if (playConfig.play_method === 'BUFFER' && !playConfig.external_audio) {
+                console.log('Autoplay buffer');
+                latency = webAudio.getTotalLatency();
+                webAudio.playBufferFrom(section.id, 0, Math.max(0, playConfig.playhead));
+            } else {
+                console.log('Autoplay HTML audio')
+                audio.playFrom(Math.max(0, playConfig.playhead));
+            }                    
+            setTimeout(startedPlaying(), latency);
         }
+    }, [playConfig, startedPlaying]);
 
-        // Clean up
-        return () => {
-            audio.pause();
-        };
-    }, [state, playConfig, startedPlaying]);
-
-
-    // Render component based on view
-    switch (state.view) {
-        case PRELOAD:           
-            return (
-                <Preload
-                    instruction={preloadMessage}
-                    duration={playConfig.ready_time}
-                    url={MEDIA_ROOT + section.url}
-                    onNext={() => {                        
-                        setView(RECOGNIZE);
-                        onPreloadReady();
+    // Render component
+    return (
+        <div>
+            <div className="circle">
+                <Circle
+                    running={running}
+                    duration={responseTime}
+                    color="white"
+                    animateCircle={playConfig.show_animation}
+                    onTick={onCircleTimerTick}
+                    onFinish={() => {
+                        // Stop audio
+                        finishedPlaying();
                     }}
                 />
-            );
-        case RECOGNIZE:            
-            return (
-                <div>
-                    <div className="circle">
-                        <Circle
-                            running={running}
+                <div className="circle-content">
+                    {playConfig.show_animation
+                        ? <ListenCircle
                             duration={responseTime}
-                            color="white"
-                            animateCircle={playConfig.show_animation}
-                            onTick={onCircleTimerTick}
-                            onFinish={() => {
-                                // Stop audio
-                                finishedPlaying();
-                            }}
+                            histogramRunning={running}
+                            countDownRunning={running}
                         />
-                        <div className="circle-content">
-                            {playConfig.show_animation
-                                ? <ListenCircle
-                                    duration={responseTime}
-                                    histogramRunning={running}
-                                    countDownRunning={running}
-                                />
-                                : <div className="stationary">
-                                    <span className="fa-solid fa-headphones fa-6x"></span>
-                                </div>
-                            }
+                        : <div className="stationary">
+                            <span className="fa-solid fa-headphones fa-6x"></span>
                         </div>
-                    </div>
-                    <div className={
-                        "aha__listen d-flex flex-column justify-content-center align-items-center " +
-                        className
                     }
-                    >
-                    {/* Instruction */}
-                    {instruction && (<div className="instruction d-flex justify-content-center align-items-center">
-                        <h3 className="text-center">{instruction}</h3>
-                    </div>)}
                 </div>
             </div>
-            )
-            default:
-                return <div>Unknown view: {state.view}</div>;
-        }
+            <div className={
+                "aha__listen d-flex flex-column justify-content-center align-items-center " +
+                className
+            }
+            >
+            {/* Instruction */}
+            {instruction && (<div className="instruction d-flex justify-content-center align-items-center">
+                <h3 className="text-center">{instruction}</h3>
+            </div>)}
+        </div>
+    </div>
+    )   
 }
 export default AutoPlay;
