+++ frontend/src/components/Preload/Preload.js
@@ -3,15 +3,17 @@ import React, { useEffect, useRef, useState } from "react";
 import ListenFeedback from "../Listen/ListenFeedback";
 import CountDown from "../CountDown/CountDown";
 import * as audio from "../../util/audio";
+import * as webAudio from "../../util/webAudio";
+import { MEDIA_ROOT } from "../../config";
 import classNames from "classnames";
 
 // Preload is an experiment screen that continues after a given time or after an audio file has been preloaded
-const Preload = ({ instruction, pageTitle, duration, url, onNext }) => {
+const Preload = ({ instruction, pageTitle, duration, sections, playConfig, onNext }) => {
     const timeHasPassed = useRef(false);
     const audioIsAvailable = useRef(false);
     const [loaderDuration, setLoaderDuration] = useState(duration);
     const [overtime, setOvertime] = useState(false);
-
+    
     const onTimePassed = () => {
         timeHasPassed.current = true;
         setLoaderDuration(0);
@@ -20,17 +22,44 @@ const Preload = ({ instruction, pageTitle, duration, url, onNext }) => {
             onNext();
         }
     };
+
     // Audio preloader
     useEffect(() => {
-        // Load audio until available
-        // Return remove listener
-        return audio.loadUntilAvailable(url, () => {
-            audioIsAvailable.current = true;
-            if (timeHasPassed.current) {
-                onNext();
+        if (playConfig.play_method === 'BUFFER' && !playConfig.external_audio) {
+
+            // Use Web-audio and preload sections in buffers            
+            sections.map((section, index) => {
+                // Clear buffers if this is the first section
+                if (index === 0) {
+                    webAudio.clearBuffers();
+                }
+                
+                // Load sections in buffer
+                // Return remove listener  
+                return webAudio.loadBuffer(section.id, MEDIA_ROOT + section.url, () => {                    
+                    if (index === (sections.length - 1)) {
+                        audioIsAvailable.current = true;
+                        if (timeHasPassed.current) {
+                            onNext();
+                        }                        
+                    }                                        
+                });
+            })
+        } else {
+            if (playConfig.external_audio) {
+                webAudio.closeWebAudio();
             }
-        });
-    }, [url, onNext]);
+            
+            // Load audio until available
+            // Return remove listener   
+            return audio.loadUntilAvailable(MEDIA_ROOT + sections[0].url, () => {
+                audioIsAvailable.current = true;
+                if (timeHasPassed.current) {
+                    onNext();
+                }
+            });            
+        }              
+    }, [sections, onNext]);
 
 
     return (
